Added choose( )
Choose takes in a list of filenames, displays these to the user and allows them to choose from those listed. It will error check and ensure the user is appropriate. It was useful to put this into a function because this method is used several times throughout the code.

As well as this addition, there have been several cleanups of the find_file function(s), including improved string formatting, better stdout messages, standardizing of format across all functions.